requirements.txt
--index-url=https://pypi.tuna.tsinghua.edu.cn/simple
Pillow==6.1.0